Remove the AWS module, move S3Writer into the Dockly namespace
