+++ lib/dockly/docker.rb
@@ -244,7 +244,7 @@ class Dockly::Docker
       if s3_bucket.nil?
         output = File.open(tar_path, 'wb')
       else
-        output = Dockly::AWS::S3Writer.new(Dockly.s3, s3_bucket, s3_object)
+        output = Dockly::S3Writer.new(Dockly.s3, s3_bucket, s3_object)
       end
 
       gzip_output = Zlib::GzipWriter.new(output)
