+++ spec/spec_helper.rb
@@ -7,8 +7,6 @@ require 'pry'
 
 Fog.mock!
 
-Dockly::AWS.aws_access_key_id = 'MOCK_KEY'
-Dockly::AWS.aws_secret_access_key = 'MOCK_SECRET'
 Dockly::Util::Logger.disable! unless ENV['ENABLE_LOGGER'] == 'true'
 Dockly.aws_region 'us-east-1'
